week4/rag_service/pyproject.toml
[project]
name = "ai-mutual-mentorship-scheme"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=0.33.0",
    "chroma-haystack>=3.4.1",
    "chromadb>=0.4",
    "haystack-ai>=2.21",
    "ollama-haystack>=5.3.0",
    "opensearch-haystack>=6.2.0",
    "pypdf>=4.0",
    "python-dotenv>=1.0",
    "requests>=2.31.0",
    "sentence-transformers>=2.6.1",
    "streamlit>=1.54.0",
    "trafilatura==2.0.0",
    "transformers>=4.46.0",
    "uvicorn>=0.41.0",
]
